wasmparser: fix component start function validation.
This commit fixes the validation of component start function parameters.

Previously, validation checked that the argument type was equal to the
parameter type. However, the types being checked with interface type references
that should not support equality checks.

The fix is to instead ensure that the argument's type is a subtype of the
parameter's type according to the subtyping rules. Additionally, the
`PartialEq` and `Eq` implementation was removed for `InterfaceTypeRef` to
prevent incorrect checks like this in the future.

Fixes #584.