wasm-smith: Hold references to the actual type in addition to the index
First of all, this just makes working with these types easier, and isn't an
issue because they are immutable and we've already decided to generate them as
is.

Second, this helps us avoid issues where an aliased compound type will reference
other types by index and that index is in a different types index space than
where it is aliased from, and therefore it is easy to look up types in the wrong
index space. An alternative would be to keep track of which index space we are
talking about, but that is harder to work with than just having a reference to
the type itself on hand.