Reduce repetition in the text format of canonical options (#628)
Currently canonical options have a textual syntax that looks like:

    (memory (core memory $name))

but personally this seems a bit wordy and repetitive to me. The "core"
part is already implied by the canonical lifting and lowering, and
otherwise the "memory" part is repeated. Alternatively I think it might
be a bit easier to read/write syntax of the form:

    (memory $name)
    (realloc $realloc)

without any loss in clarity. Additionally though this would still keep
the ability to reference and instance along the lines of:

    (memory $libc "memory")