wasm-smith: fix subtraction with overflow.
This commit fixes a subtraction with overflow that occurs in
`arbitrary_outer_core_type_alias` when there is exactly one function type in
the parent scope and no function types in the local scope.

The problem was that the inclusive maximum was subtracted from twice.

The fix is to only subtract for the inclusive maximum once.