wasm-smith: limit component function type parameter counts.
This commit limits `wasm-smith` to 16 parameters so that the `memory` option is
not required for any canonical functions that might use the function type.

As there's currently a 1:1 between parameter type and core wasm value type for
component function signatures, more than 16 parameters will necessitate
indirect passing of the parameters via memory; when this occurs, component
validation will require the `memory` option to be present.

However, `wasm-smith` doesn't yet set any options for the canonical functions
it generates.